requirements.txt
Django==2.2
django-bootstrap4==0.0.7
django-bootstrap3==10.0.1
django-filter==2.0.0
django-tables2==2.0.3
django-debug-toolbar==1.9.1
django-markup==1.3
tablib<0.11.99
django-markup[all-filter-dependencies]
djangorestframework==3.9.0
docutils==0.14
docopt==0.6.2
drf-tracking==1.5.0
drfdocs==0.0.11
Markdown==3.0.1
python-dateutil==2.7.5
pytz==2018.9
requests==2.21.0
six==1.12.0
urllib3==1.24.1
whitenoise==4.1.2
yarg==0.1.9
gunicorn
